Fix errors while implementing tab system. not perfect yet!
