+++ Yttrium/WebViewPage.xaml
@@ -1,5 +1,5 @@
-﻿<Page
-    x:Class="Yttrium.WebViewPage"
+﻿<winui:TabViewItem
+    x:Class="Yttrium.WebViewTab"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -18,4 +18,4 @@
                 Source="https://google.com/"
                 x:FieldModifier="public"/>
     </Frame>
-</Page>
+</winui:TabViewItem>
